% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sammyR-package.R
\docType{package}
\name{sammyR-package}
\alias{sammyR}
\alias{sammyR-package}
\title{sammyR: Analysis of Sequential Analysis of MacroMolecules accessibilitY sequencing (SAMMY-seq) Data}
\description{
Provides functions to analyse Sequential Analysis of MacroMolecules accessibilitY sequencing (SAMMY-seq) data, including (a) chromatin compartmentalization analysis to identify A/B compartments across chromosomes via CALDER2 and (b) identification of differentially enriched solubility regions as described in Wang et al. (2025).
}
\author{
\strong{Maintainer}: Margherita Mutarelli \email{margherita.mutarelli@cnr.it}

Authors:
\itemize{
  \item Margherita Mutarelli \email{margherita.mutarelli@cnr.it}
}

}
\keyword{internal}
