% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartments.R
\name{subanno_maker}
\alias{subanno_maker}
\title{Build a Gviz annotation track from a subcompartment GRanges}
\usage{
subanno_maker(subcomp_gr, patient)
}
\arguments{
\item{subcomp_gr}{Subcompartment \code{GRanges} (from \code{subgr_extractor}).}

\item{patient}{Sample name used as the track name.}
}
\value{
A Gviz \code{AnnotationTrack}.
}
\description{
Build a Gviz annotation track from a subcompartment GRanges
}
\keyword{internal}
